+++ CHANGES.rst
@@ -5,13 +5,18 @@ Changes
 -----
 
 - Add database migrations framework and automatic migrations applying.
+- External access links for styles, webmaps (TMS) and feature layers (MVT).
+- Experimental WFS client and raster mosaic which is disabled by default.
 - Add support of 1.1.0 version in WFS server implementation.
 - Improved handling of NODATA values in raster layer and raster style.
 - Compression level of PNG images is set to 3 which is much faster.
+- Performance improvements and better concurency for tile cache.
 - New "CSV for Microsoft Excel" export format for better Excel compatibility.
 - Fix infinite wait of database lock, including during vector layer deletion.
 - Improved handling of invalid JSON bodies in RESP API, now correct error
   message returns.
+- Vector layer export to MapInfo MIF/MID format.
+- Vector layer export to Panorama SXF format.
 
 3.6.0
 -----